requirements.txt
tkinter
web3
requests
mnemonic
cryptography
pillow
eth-account

